+++ packages/assignment-7/src/constants.ts
@@ -5,3 +5,7 @@ export const notificationOptions = [
   { value: 120, label: "2시간 전" },
   { value: 1440, label: "1일 전" },
 ];
+
+export const categories = ["업무", "개인", "가족", "기타"];
+
+export const weekDays = ["일", "월", "화", "수", "목", "금", "토"];
